cast input to int64 when input offset is used to avoid overflow due to table concat

+++ distributed_embeddings/python/layers/dist_model_parallel.py
@@ -436,7 +436,7 @@ class DistributedEmbedding(tf.keras.layers.Layer):
 
     # offset inputs
     inputs = [
-        inp if offset is None else inp + tf.cast(offset, inp.dtype)
+        inp if offset is None else tf.cast(inp, tf.int64) + offset
         for inp, offset in zip(inputs, self.offsets)
     ]
     # do embedding
